zombienet setup
