+++ examples/parachain-example/src/lib.rs
@@ -173,6 +173,10 @@
 //!
 //! All configs we constructed, `StatemintConfig`, `StatemintConfig2` and `StatemintConfig3` should behave in the same way.
 //! All three ways are valid for constructing a config. Choose one depending on your use case.
+//! 
+//! # 3 Try the examples
+//! We have two examples of interacting with the parachain, that you can try. You find them in the bin directory of this crate. Please check them out.
+//! 
 //!
 use codec::Encode;
 use futures::StreamExt;
@@ -362,55 +366,4 @@ impl Config for StatemintConfig3 {
     type Header = <PolkadotConfig as Config>::Header;
     // this is the only difference:
     type ExtrinsicParams = <SubstrateConfig as Config>::ExtrinsicParams;
-}
-
-/// In this example you can just switch out `StatemintConfig` for `StatemintConfig2` or `StatemintConfig3` and the behavior should be the same.
-#[tokio::main]
-async fn main() -> Result<(), Box<dyn std::error::Error>> {
-    let rpc_endpoint = "wss://polkadot-asset-hub-rpc.polkadot.io:443";
-
-    // here we use the config:
-    let api = OnlineClient::<StatemintConfig>::from_url(rpc_endpoint).await?;
-
-    // Subscribe to all finalized blocks:
-    let mut blocks_sub = api.blocks().subscribe_finalized().await?;
-
-    // For each block, print a bunch of information about it:
-    while let Some(block) = blocks_sub.next().await {
-        let block = block?;
-
-        let block_number = block.header().number;
-        let block_hash = block.hash();
-
-        println!("Block #{block_number}:");
-        println!("  Hash: {block_hash}");
-        println!("  Extrinsics:");
-
-        let body = block.body().await?;
-        for ext in body.extrinsics().iter() {
-            let ext = ext?;
-            let idx = ext.index();
-            let events = ext.events().await?;
-            // here we make use of the generated metadata code:
-            let decoded_ext = ext.as_root_extrinsic::<metadata::statemint::Call>();
-
-            println!("    Extrinsic #{idx}:");
-            println!("      Bytes: {}", ext.bytes().len());
-            println!("      Decoded: {decoded_ext:?}");
-            println!("      Events:");
-
-            for evt in events.iter() {
-                let evt = evt?;
-
-                let pallet_name = evt.pallet_name();
-                let event_name = evt.variant_name();
-                let event_values = evt.field_values()?;
-
-                println!("        {pallet_name}_{event_name}");
-                println!("          {}", event_values);
-            }
-        }
-    }
-
-    Ok(())
 }
